Cleaned env, move build to ci
