+++ .github/workflows/build-extension.yml
@@ -1,4 +1,4 @@
-name: Lint & Build (extension)
+name: Typecheck (extension)
 
 on:
   push:
@@ -13,20 +13,6 @@ permissions:
   pull-requests: write
 
 jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Setup Node Dependencies
-        uses: ./.github/actions/setup-node-deps
-        with:
-          node-version: "22.22.0"
-          build-sdks: "true"
-          build-sparkle: "true"
-      - working-directory: extension
-        run: npm run package:chrome:production
-        env:
-          DATADOG_CLIENT_TOKEN: ${{ secrets.DD_CLIENT_TOKEN }}
   tsgo:
     runs-on: ubuntu-latest
     steps:
