+++ extension/.env.development
@@ -1,7 +1,2 @@
 NODE_ENV=development
-WORKOS_CLIENT_ID=client_01JGCT55EJ328PJ8MSCVSKVDKE
-WORKOS_DOMAIN=api.workos.com
-DEFAULT_DUST_API_DOMAIN="http://localhost:3000"
-WORKOS_CLAIM_NAMESPACE="https://dust.tt/"
-DUST_US_URL="http://localhost:3000"
 FRONT_EXTENSION_URL="http://localhost:3012"
