+++ extension/.env.production
@@ -1,10 +1,2 @@
 NODE_ENV=production
-WORKOS_CLIENT_ID=client_01JGCT55T7FVDG9XF74925R1KT
-WORKOS_DOMAIN=api.workos.com
-DEFAULT_DUST_API_DOMAIN="https://dust.tt"
-WORKOS_CLAIM_NAMESPACE="https://dust.tt/"
-DUST_US_URL="https://dust.tt"
-DUST_EU_URL="https://eu.dust.tt"
 FRONT_EXTENSION_URL="https://front-ext.dust.tt"
-NEXT_PUBLIC_DUST_APP_URL="https://app.dust.tt"
-VIZ_PUBLIC_URL="https://viz.dust.tt"
