+++ extension/platforms/chrome/main.tsx
@@ -17,7 +17,7 @@ if (process.env.DATADOG_CLIENT_TOKEN) {
     site: "datadoghq.eu",
     service: "dust-chrome-extension",
     env: process.env.DATADOG_ENV,
-    version: process.env.VERSION,
+    version: process.env.DUST_EXTENSION_VERSION,
     forwardConsoleLogs: ["error"],
     forwardErrorsToLogs: true,
     sessionSampleRate: 100,
