+++ extension/platforms/front/webpack.config.ts
@@ -1,6 +1,7 @@
 import type { Environment } from "@extension/config/env";
 import { execSync } from "child_process";
 import Dotenv from "dotenv-webpack";
+import fs from "fs";
 import HtmlWebpackPlugin from "html-webpack-plugin";
 import path from "path";
 import TerserPlugin from "terser-webpack-plugin";
@@ -18,7 +19,10 @@ const getCommitHash = () => {
 
 export const getConfig = ({ env }: { env: Environment }) => {
   const isDevelopment = env === "development";
-
+  const packageJson = JSON.parse(
+    fs.readFileSync(path.resolve(__dirname, "../../package.json"), "utf8")
+  );
+  const version = packageJson.version;
   return {
     mode: isDevelopment ? "development" : "production",
     entry: path.resolve(__dirname, "./main.tsx"),
@@ -100,18 +104,24 @@ export const getConfig = ({ env }: { env: Environment }) => {
       }),
 
       new webpack.EnvironmentPlugin({
-        COMMIT_HASH: getCommitHash(),
+        BUILD_DATE: process.env.COMMIT_HASH || Math.floor(Date.now() / 1000),
+        COMMIT_HASH: process.env.COMMIT_HASH || getCommitHash(),
         DATADOG_CLIENT_TOKEN: process.env.DATADOG_CLIENT_TOKEN || "",
         DATADOG_ENV: isDevelopment ? "dev" : "prod",
-        NEXT_PUBLIC_VIRTUOSO_LICENSE_KEY:
-          process.env.NEXT_PUBLIC_VIRTUOSO_LICENSE_KEY || "",
+        DUST_EU_URL: process.env.DUST_EU_URL || "",
+        DUST_EXTENSION_VERSION: `front-${version}`,
+        DUST_US_URL: process.env.DUST_US_URL || "",
+        FRONT_EXTENSION_URL: process.env.FRONT_EXTENSION_URL || "",
+        NEXT_PUBLIC_DUST_APP_URL: process.env.NEXT_PUBLIC_DUST_APP_URL || "",
+        NEXT_PUBLIC_NOVU_API_URL: process.env.NEXT_PUBLIC_NOVU_API_URL || "",
         NEXT_PUBLIC_NOVU_APPLICATION_IDENTIFIER:
           process.env.NEXT_PUBLIC_NOVU_APPLICATION_IDENTIFIER || "",
-        NEXT_PUBLIC_NOVU_API_URL: process.env.NEXT_PUBLIC_NOVU_API_URL || "",
         NEXT_PUBLIC_NOVU_WEBSOCKET_API_URL:
           process.env.NEXT_PUBLIC_NOVU_WEBSOCKET_API_URL || "",
-        NEXT_PUBLIC_DUST_APP_URL: process.env.NEXT_PUBLIC_DUST_APP_URL || "",
+        NEXT_PUBLIC_VIRTUOSO_LICENSE_KEY:
+          process.env.NEXT_PUBLIC_VIRTUOSO_LICENSE_KEY || "",
         VIZ_PUBLIC_URL: process.env.VIZ_PUBLIC_URL || "",
+        WORKOS_CLIENT_ID: process.env.WORKOS_CLIENT_ID || "",
       }),
       new Dotenv({
         path: isDevelopment
