+++ extension/shared/lib/config.ts
@@ -1,10 +1,4 @@
 export const WORKOS_CLIENT_ID = process.env.WORKOS_CLIENT_ID ?? "";
-export const WORKOS_DOMAIN =
-  process.env.WORKOS_DOMAIN ?? "https://api.workos.com";
-export const WORKOS_CLAIM_NAMESPACE = process.env.WORKOS_CLAIM_NAMESPACE ?? "";
-
 export const DUST_US_URL = process.env.DUST_US_URL ?? "";
 export const DUST_EU_URL = process.env.DUST_EU_URL ?? "";
-export const DEFAULT_DUST_API_DOMAIN =
-  process.env.DEFAULT_DUST_API_DOMAIN ?? "";
 export const FRONT_EXTENSION_URL = process.env.FRONT_EXTENSION_URL ?? "";
